services/ingestion/elections/pyproject.toml
[project]
name = "civic-ingest-elections"
version = "0.0.0"
description = "V2 elections adapter — official CEC national election results."
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.9",
  "structlog>=24.4",
  "PyYAML>=6",
  "civic-ingest",
  "civic-archival",
  "civic-clients",
  "civic-ontology",
]

[dependency-groups]
dev = ["pytest>=8"]

[tool.uv.sources]
civic-ingest = { workspace = true }
civic-archival = { workspace = true }
civic-clients = { workspace = true }
civic-ontology = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/civic_ingest_elections"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
